Allows preview of MD files
With this you need to manually install the program and use it on a trial
an error basis. The happy-path works but I don't think the CSS trick is
working and I am not really worried about it right now.

[ ] Make it public
[ ] Create the https://git.io URL (it doens't exist, now)
[ ] Make the HTML generated use the CSS (bug)
[x] Allow specification of other formats (.txt, .rtf)
[x] Allow option to prefer opening the preview in a browser
[x] Document variables which affect behavior

Although there is much left to do, this can be used right after the
installation: it will install `pandoc` using `brew` if it is not
present. It also lazy-downloads a CSS that mimics GitHub rendering of
markdown.