Enables Linux users to have the same OSX experience
Executing "preview-markdown" inside a repository root folder and without
any arguments should open the browser with the preview of the 'README.md'.
On Linux, just the HTML file path is output.

This installs pandoc on Debian based systems, if necessary, and uses
"sensible-browser" to open the browser.

CSS Bug
-------

On OSX I don't ever remember the CSS file working, but on Ubuntu it
does.
In the same manner, the auto text encoding detection on the browser
works on Ubuntu while it fails (to favor Unicode) on OSX.